move handleLegacy back into handleMessage
