+++ explorer/src/server/handleMessage.ts
@@ -1,11 +1,26 @@
 import rpcServer from './rpcServer'
-import handleLegacy from './rpcMethods/legacy'
+import { logger } from '../logging'
+import { getDb } from '../database'
+import { fromString, saveJobRunTree } from '../entity/JobRun'
 
-export type messageContext = {
+export type serverContext = {
   chainlinkNodeId: number
 }
 
-const handleJSONRCP = async (request: string, context: messageContext) => {
+const handleLegacy = async (json: string, context: serverContext) => {
+  try {
+    const db = await getDb()
+    const jobRun = fromString(json)
+    jobRun.chainlinkNodeId = context.chainlinkNodeId
+    await saveJobRunTree(db, jobRun)
+    return { status: 201 }
+  } catch (e) {
+    logger.error(e)
+    return { status: 422 }
+  }
+}
+
+const handleJSONRCP = async (request: string, context: serverContext) => {
   return await new Promise((resolve, reject) => {
     // @ts-ignore - broken typing for server.call - should be able to accept 3 arguments
     // https://github.com/tedeh/jayson#server-context
@@ -19,7 +34,7 @@ const handleJSONRCP = async (request: string, context: messageContext) => {
 
 export const handleMessage = async (
   message: string,
-  context: messageContext,
+  context: serverContext,
 ) => {
   if (message.includes('jsonrpc')) {
     return await handleJSONRCP(message, context)
